requirements.txt
# HuntForge Core Dependencies
pyyaml==6.0.1
rich==13.7.1
loguru==0.7.2
requests==2.31.0
python-dotenv==1.0.1

# HuntForge Dashboard
Flask==3.0.3

# AI Generation SDKs
google-generativeai==0.5.0
